backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas>=2.2.0
openpyxl==3.1.2
selenium==4.16.0
mysql-connector-python==8.2.0
